+++ Opcodes/ugens9.c
@@ -115,7 +115,7 @@ static int cvset(CSOUND *csound, CONVOLVE *p)
               + (p->nchanls > 1 ? (Hlenpadded + 2) : 0));
     if (p->auxch.auxp == NULL || p->auxch.size < siz*sizeof(MYFLT)) {
       /* if no buffers yet, alloc now */
-      csound->AuxAlloc(csound, (long) siz*sizeof(MYFLT), &p->auxch);
+      csound->AuxAlloc(csound, (size_t) siz*sizeof(MYFLT), &p->auxch);
       fltp = (MYFLT *) p->auxch.auxp;
       p->fftbuf = fltp;   fltp += (Hlenpadded + 2); /* and insert addresses */
       p->olap = fltp;     fltp += p->nchanls*(Hlen - 1);
